Completed the constructor for class Segment

+++ System/src/partition/segment.hh
@@ -45,9 +45,9 @@ class Segment
 		/* FilePartition the Segment belongs to */
 		FilePartition& _partition;
 		/* FilePartition the Segment belongs to */
-		uint32_t _maxSize;
+		uint16_t _maxSize;
 		/* Position in Partition */
-		uint _index;
+		uint32_t _index;
         /* Page Header of Segment */
 		segment_page_header_t _header;
 };
